refactor: refactor storage management to common module and remove popup storage
